# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Math computation API with history
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: math
    description: Math computation operations
  - name: history
    description: Computation history operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /math/compute:
    post:
      operationId: computeMath
      tags: [math]
      summary: Compute a math expression
      description: Evaluates a mathematical expression and returns step-by-step results. Supports arithmetic, algebra (factor, expand, solve, differentiate, integrate, limit, simplify), and polynomial operations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeRequest"
      responses:
        "200":
          description: Computation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResult"
        "400":
          description: Invalid expression or unknown operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: Expression could not be parsed or evaluated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /math/operations:
    get:
      operationId: listOperations
      tags: [math]
      summary: List supported operations
      description: Returns all supported math operations with descriptions and examples
      responses:
        "200":
          description: List of supported operations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: List computation history
      description: Returns recent computation history, newest first
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return (1–100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: operation
          in: query
          description: Filter by operation type
          schema:
            type: string
      responses:
        "200":
          description: Computation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryList"
    delete:
      operationId: clearHistory
      tags: [history]
      summary: Clear all computation history
      description: Deletes all stored computation history records
      responses:
        "204":
          description: History cleared successfully
  /history/{id}:
    get:
      operationId: getHistoryEntry
      tags: [history]
      summary: Get a single history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteHistoryEntry
      tags: [history]
      summary: Delete a single history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Entry deleted successfully
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error

    ComputeRequest:
      type: object
      properties:
        expression:
          type: string
          description: The mathematical expression to evaluate (e.g. "x^2 + 2x + 1", "3 + 4 * 2")
          minLength: 1
          maxLength: 1000
        operation:
          type: string
          description: "The operation to perform: factor, expand, solve, diff, integrate, limit, simplify, evaluate. If omitted, auto-detected from the expression."
          enum:
            - factor
            - expand
            - solve
            - diff
            - integrate
            - limit
            - simplify
            - evaluate
        variable:
          type: ["string", "null"]
          description: Variable to use for symbolic operations (default 'x')
          default: "x"
        limitPoint:
          type: ["number", "null"]
          description: Point at which to evaluate a limit (only used when operation='limit'). Default is 0.
      required:
        - expression

    ComputeStep:
      type: object
      properties:
        step:
          type: integer
          description: Step number (1-indexed)
        label:
          type: string
          description: Human-readable description of this step
        expression:
          type: string
          description: The expression at this step (LaTeX or plain text)
      required:
        - step
        - label
        - expression

    ComputeResult:
      type: object
      properties:
        expression:
          type: string
          description: The original input expression
        operation:
          type: string
          description: The operation that was performed
        result:
          type: string
          description: The final computed result as a string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ComputeStep"
          description: Step-by-step breakdown of the computation
        isNumeric:
          type: boolean
          description: True if the result is a pure numeric value
        numericValue:
          type: ["number", "null"]
          description: Numeric value of the result if applicable
        historyId:
          type: integer
          description: ID of the saved history entry
      required:
        - expression
        - operation
        - result
        - steps
        - isNumeric

    OperationInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        examples:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - examples

    OperationList:
      type: object
      properties:
        operations:
          type: array
          items:
            $ref: "#/components/schemas/OperationInfo"
      required:
        - operations

    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        expression:
          type: string
        operation:
          type: string
        result:
          type: string
        steps:
          type: string
          description: JSON-encoded steps array
        isNumeric:
          type: boolean
        numericValue:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - expression
        - operation
        - result
        - steps
        - isNumeric
        - createdAt

    HistoryList:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        total:
          type: integer
      required:
        - entries
        - total